add docker compose file for the database
This allows us to easily run a PostgreSQL DB with a simple UI for ease of debugging.

In the future, adminer will disappear and the backend service will be included in the docker file so you can run everything with one docker compose command.

Secrets/passwords shouldn't be hardcoded here either, but it's all for the sake of getting to a minimum setup fast now.