add tests for bookmarks handler getBookmarks endpoint
This might look ugly to someone used to Supertest in terms of readability, but ultimately I decided against having to add that much bloat to the project + having to run integration tests as opposed to more localized unit tests which I prefer.

I think whenever we have an openAPI spec of the service, we might be able to automate some e2e tests to complement those unit ones and that should suffice.

The service layer is pretty slim by now and the store is pretty Knex dependent so for now I feel like this is a good compromise.

I might need to find a way to not have files full of @ts-ignore comments but I wanted to be able to get fast to a functional state without having to fight with Typescript for my tests.