add tests for addBookmark endpoint
This also triggered a refinement of responses when errors occur. Now we wrap everything in an `error` property, so we can both return a human-friendly message and an error code that can be used by clients to easily match cases and implement different behaviours per error type.

The `getBookmarks` endpoint will receive the same treatment in an upcoming commit.